[SPARK-19869][SQL]move table related ddl from ddl.scala to tables.scala
